+++ cmd/cycloid/pipelines/list.go
@@ -1,27 +1,56 @@
 package pipelines
 
 import (
-	"fmt"
+	"os"
 
-	"github.com/cycloidio/cycloid-cli/cmd/cycloid/internal"
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
+
+	"github.com/cycloidio/cycloid-cli/cmd/cycloid/common"
+	"github.com/cycloidio/cycloid-cli/cmd/cycloid/internal"
+	"github.com/cycloidio/cycloid-cli/cmd/cycloid/middleware"
+	"github.com/cycloidio/cycloid-cli/printer"
+	"github.com/cycloidio/cycloid-cli/printer/factory"
 )
 
 func NewListCommand() *cobra.Command {
 	var cmd = &cobra.Command{
 		Use:     "list",
-		Hidden:  true,
-		Short:   "not implemented yet",
-		Long:    `not implemented yet`,
+		Short:   "list all pipelines of the organization",
 		PreRunE: internal.CheckAPIAndCLIVersion,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			fmt.Println("not implemented yet")
-			return nil
-		},
+		RunE:    list,
 	}
+
 	return cmd
 }
 
-// /organizations/{organization_canonical}/pipelines
-// get: getPipelines
-// Get all the pipelines that the authenticated user has access to.
+func list(cmd *cobra.Command, args []string) error {
+	api := common.NewAPI()
+	m := middleware.NewMiddleware(api)
+
+	org, err := cmd.Flags().GetString("org")
+	if err != nil {
+		return err
+	}
+	output, err := cmd.Flags().GetString("output")
+	if err != nil {
+		return errors.Wrap(err, "unable to get output flag")
+	}
+
+	jobs, err := m.ListPipelines(org)
+	if err != nil {
+		return errors.Wrap(err, "unable to list pipelines")
+	}
+
+	// fetch the printer from the factory
+	p, err := factory.GetPrinter(output)
+	if err != nil {
+		return errors.Wrap(err, "unable to get printer")
+	}
+
+	// print the result on the standard output
+	if err := p.Print(jobs, printer.Options{}, os.Stdout); err != nil {
+		return errors.Wrap(err, "unable to print result")
+	}
+	return nil
+}